requirements.txt
Flask
opencv-python
numpy
werkzeug
pymupdf
Pillow
